# ports/rp2: Mark gc_heap NOLOAD for faster boot.