@Gadgetoid 2022-08-05
@jimmo I think:

> It should be pretty straightforward to put it in a different section (e.g. .noinit), or better do what we do on stm32 where it's just defined entirely in the linker config as a start and end symbol used directly by gc_init().

Might depend upon: https://github.com/micropython/micropython/pull/8761