@dpgeorge 2022-08-09
It would definitely be good to improve rp2 start up speed, and this looks like a nice and simple thing to do to begin with.