@Gadgetoid 2022-08-10
This has made me wonder what `uninitialized_data` is actually for, since I can't find any instance of it being used or even a passing reference to it at all.

I've raised an issue against Pico SDK since there's either something amiss or I'm very confused - 
https://github.com/raspberrypi/pico-sdk/issues/966