@Gadgetoid 2022-08-10
Just to go a little further with my `unitialized_bss` is correct hypothesis (I ... can't really speak authoritatively on this stuff so I'm relying on Google and experimentation)-

Normal build:

```
   text	   data	    bss	    dec	    hex	filename
 702976	    124	 205064	 908164	  ddb84	firmware.elf
 ```

With `unitialized_bss`:

```
   text	   data	    bss	    dec	    hex	filename
 702976	    124	 205064	 908164	  ddb84	firmware.elf
```

With `unitialized_data`:

```
   text	   data	    bss	    dec	    hex	filename
 702976	 170108	  35080	 908164	  ddb84	firmware.elf
 ```

The linker script `NOLOAD` directive is - I believe - what makes the difference here and pending some clarification for why `unitialized_data` even exists and what it's intended to achieve I think it's worth an `unitialized_bss`.

In terms of the final result- you were correct, simply marking `gc_heap` as `uninitialized_data` prevented it from being loaded.