@dpgeorge 2022-08-11
Parts of the VM/runtime are relocated to RAM at start up, and this could probably also benefit from improvements.  Eg only copy this data after switching to a higher clock frequency.