@Gadgetoid 2022-08-11
> Parts of the VM/runtime are relocated to RAM at start up, and this could probably also benefit from improvements. Eg only copy this data after switching to a higher clock frequency.

This is the main thrust of my speedup hacks which are applied as a patch to `crt0.S` and `runtime.c` before building in our CI.

The Pico SDK maintainers have shown some interest in making this easier, but I don't know how imminent that will be.

You can currently use `SKIP_PICO_STANDARD_LINK` and `SKIP_PICO_RUNTIME` flags to remove those altogether and supply your own, but that's pretty drastic.

There's some discussion and links out to relevant issues here - https://github.com/raspberrypi/pico-sdk/issues/959

As I mentioned above, this could potentially bring the stock Pico MicroPython startup to ~30-50ms. A dramatic improvement over even this change, since all copies to RAM (VM, runtime and initialized data) are sped up.

Definitely something to keep in mind, since I think early clock setup - via whatever solution Pico SDK canonicalizes upon - would likely be the easiest way to accomplish what you mention.