@dpgeorge 2022-08-11
Let's keep an eye on that pico-sdk ticket.  There are definitely gains to be made here in terms of improving startup time.